代码日常更新

+++ anyoujia-web/src/main/java/cn/anyoufang/service/impl/LockMemberServiceImpl.java
@@ -1,22 +1,26 @@
 package cn.anyoufang.service.impl;
 
-import cn.anyoufang.entity.SpMemberRelation;
-import cn.anyoufang.entity.SpMemberRelationExample;
+import cn.anyoufang.entity.*;
 import cn.anyoufang.exception.LockException;
+import cn.anyoufang.mapper.SpAdminLockMapper;
+import cn.anyoufang.mapper.SpLockFingerMapper;
+import cn.anyoufang.mapper.SpLockPasswordMapper;
 import cn.anyoufang.mapper.SpMemberRelationMapper;
 import cn.anyoufang.service.LockMemberService;
 import cn.anyoufang.service.LoginService;
 import cn.anyoufang.utils.DateUtil;
-import cn.anyoufang.utils.JsonUtils;
 import cn.anyoufang.utils.Md5Utils;
 import cn.anyoufang.utils.SimulateGetAndPostUtil;
+import cn.anyoufang.utils.selfdefine.CommonUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.transaction.interceptor.TransactionAspectSupport;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.List;
 
 /**
  * @author daiping
@@ -27,6 +31,15 @@ public class LockMemberServiceImpl implements LockMemberService {
 
     @Autowired
     private SpMemberRelationMapper relationMapper;
+
+    @Autowired
+    private SpAdminLockMapper adminLockMapper;
+
+    @Autowired
+    private SpLockPasswordMapper passwordMapper;
+
+    @Autowired
+    private SpLockFingerMapper fingerMapper;
 
     @Autowired
     private LoginService  loginService;
@@ -115,21 +128,72 @@ public class LockMemberServiceImpl implements LockMemberService {
     @Transactional(rollbackFor = RuntimeException.class)
     public boolean delUser(String locksn,int userid) {
         try{
-            int index =  relationMapper.deleteByPrimaryKey(userid);
-            if(index == 1) {
-                String res =  delLockPwd(locksn,userid);
-                Map<String,Object> data = JsonUtils.jsonToMap(res);
-                Integer status = (Integer)data.get("code");
-                if(status == 200) {
+               SpMemberRelation relationUser =  relationMapper.selectByPrimaryKey(userid);
+               boolean isDelete = true;
+               if(relationUser != null) {
+                   String phone = relationUser.getPhone();
+                   SpMember member =  loginService.getUserByAccount(phone);
+                   int memberId =  member.getUid();
+                   if(relationUser.getLockpwdauth()){
+                       if(relationUser.getSetedlockpwd()) {
+                           SpLockPasswordExample example = new SpLockPasswordExample();
+                           SpLockPasswordExample.Criteria criteria =  example.createCriteria();
+                           criteria.andLocksnEqualTo(locksn).andMemberidEqualTo(memberId);
+                           List<SpLockPassword> lockPasswords =  passwordMapper.selectByExample(example);
+                          if(lockPasswords.size() >0) {
+                             SpLockPassword lockPassword =  lockPasswords.get(0);
+                             int seqPwdid = lockPassword.getPwdid();
+                              String res = delLockPwd(locksn,seqPwdid);
+                              if(!CommonUtil.successResponse(res)) {
+                                  isDelete =false;
+                              }else {
+                                  passwordMapper.deleteByPrimaryKey(seqPwdid);
+                              }
+                          }
+
+                       }
+                   }
+                   if(relationUser.getFingerpwdauth()) {
+                       if(relationUser.getSetedlockfinger()) {
+                           List<SpLockFinger> fingers =  getFingerListInternal(locksn,memberId);
+                           if(fingers.size() >0) {
+                               SpLockFinger finger = fingers.get(0);
+                               int seqFingerid = finger.getId();
+                               String res = delLockPwd(locksn,seqFingerid);
+                               if(!CommonUtil.successResponse(res)) {
+                                   isDelete = false;
+                               }else {
+                                   fingerMapper.deleteByPrimaryKey(seqFingerid);
+                               }
+                           }
+                       }
+                   }
+               }
+
+                if(isDelete) {
+                    relationMapper.deleteByPrimaryKey(userid);
                     return true;
                 }
-            }
         }catch (Exception e) {
             TransactionAspectSupport.currentTransactionStatus().setRollbackOnly();
             return false;
         }
         return false;
     }
+
+    /**
+     * 提取重复代码，查询获取用户指纹列表
+     * @param locksn
+     * @param memberId
+     * @return
+     */
+    private List<SpLockFinger> getFingerListInternal(String locksn,int memberId) {
+        SpLockFingerExample example = new SpLockFingerExample();
+        SpLockFingerExample.Criteria criteria = example.createCriteria();
+        criteria.andMemberidEqualTo(memberId).andLocksnEqualTo(locksn);
+        return fingerMapper.selectByExample(example);
+    }
+
 
     /**
      * 删除锁密码/指纹/IC卡用户信息
@@ -148,7 +212,7 @@ public class LockMemberServiceImpl implements LockMemberService {
     }
 
     /**
-     * 更新租客密码及指纹权限
+     * 更新租客过期时间
      * @param userid 成员id
      * @param locksn
      * @return
@@ -163,6 +227,64 @@ public class LockMemberServiceImpl implements LockMemberService {
         if(updatedLine == 1) {
             return true;
         }
+        return false;
+    }
+
+    /**
+     * 分别删除单个手指指纹
+     * @param seqid
+     * @param locksn
+     * @return
+     */
+    @Override
+    public boolean deleteFingerAccordent(int seqid, String locksn) {
+        String res = delLockPwd(locksn,seqid);
+        return CommonUtil.successResponse(res);
+    }
+
+    /**
+     * 获取用户指纹列表
+     * @param memberid
+     * @param locksn
+     * @return
+     */
+    @Override
+    public List<SpLockFinger> getFingerList(int memberid, String locksn) {
+      return getFingerListInternal(locksn,memberid);
+    }
+
+    /**
+     * 用于前台判断是否已经设置锁永久密码
+     * @param locksn
+     * @param phone
+     * @param memberid
+     * @return
+     */
+    @Override
+    public boolean isSetLockPwdForever(String locksn, String phone,int memberid) {
+        SpAdminLockExample example = new SpAdminLockExample();
+        SpAdminLockExample.Criteria criteria = example.createCriteria();
+        criteria.andAdminidEqualTo(memberid);
+        List<SpAdminLock> list = adminLockMapper.selectByExample(example);
+        if(list.size() == 0) {
+            SpMemberRelationExample example1 = new SpMemberRelationExample();
+           SpMemberRelationExample.Criteria criteria1 =  example1.createCriteria();
+           criteria1.andPhoneEqualTo(phone).andLocksnEqualTo(locksn);
+            List<SpMemberRelation> list1 = relationMapper.selectByExample(example1);
+            if(list1.size() >0) {
+                SpMemberRelation relation = list1.get(0);
+                boolean isSeted = relation.getSetedlockpwd();
+                if(isSeted) {
+                    return true;
+                }
+                return false;
+            }
+        }
+        SpAdminLock adminLock = list.get(0);
+        boolean isseted =  adminLock.getSetedlockpwd();
+        if(isseted) {
+           return true;
+       }
         return false;
     }
 
@@ -193,16 +315,21 @@ public class LockMemberServiceImpl implements LockMemberService {
             SpMemberRelation user = iterator.next();
             String phone =  user.getPhone();
             int end = user.getEndtime();
+            String usertype = user.getUsertype();
             //usertype 3表示租户，只有租户有过期时间，当租户30天内未注册系统时，则过期并删除
-            if(end != 0 &&!loginService.checkAccount(phone)) {
-                if ("3".equals(user.getUsertype())) {
+            if("3".equals(usertype)) {
+                if (!loginService.checkAccount1(phone)) {
                     if (DateUtil.isNotExpired(end,deadline)) {
                         validMembers.add(user);
                     } else {
                         invalidMembers.add(user);
                     }
                 }
+
+            }else {
+                validMembers.add(user);
             }
+
         }
         if(invalidMembers.size()>0) {
             SpMemberRelationExample example = new SpMemberRelationExample();
@@ -250,9 +377,9 @@ public class LockMemberServiceImpl implements LockMemberService {
 
         }else {
             if(status == 0) {
-                user.setFingerpwdauth(false);
+                user.setLockpwdauth(false);
             }else {
-                user.setFingerpwdauth(true);
+                user.setLockpwdauth(true);
             }
         }
